Projekt_UAIM/backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Cors
psycopg2-binary
Werkzeug
PyJWT==2.7.0 
